Move dialog box to `src/gui`.
